Fix filename column width and attributes
Display the basename instead of the whole path for each file when
the filename column is displayed. Make the column the same width all
the time, taking the maximum basename width of all loaded files and
padding with spaces.

Also fix the attributes when the filename column is displayed.  Insert
the correct filename-matching color for as many characters as we add for
the filename column.  Use the full filename when looking up the color
because we don't want files in different paths to have the same color,
and we don't want the color to change based on whether we are displaying
the base or the whole filename.